requirements.txt
kubernetes>=23.3.0
requests>=2.25.1
tabulate>=0.8.9
packaging>=21.3
python-dateutil>=2.8.2
PyYAML>=6.0
